This resolves #12
Implemented a block test to check if one or more block expressions (regular expressions) are defined before the build routine executes. Overlapping blocks are not supported. If a block sequence is found, the text within the sequence is treated and compared as one word. Note that any HTML tags within the sequence will be included in the word.

Updated the example to reflect the new functionality.